Factor out more stuff from ElementTrait
