+++ web/modules/custom/server_general/src/ElementTrait/CardTrait.php
@@ -1,0 +1,97 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\server_general\ElementTrait;
+
+use Drupal\server_general\ElementLayoutTrait;
+
+/**
+ * Helper methods for rendering Card elements.
+ */
+trait CardTrait {
+
+  use ElementLayoutTrait;
+
+  /**
+   * Wrap multiple cards with a grid.
+   *
+   * @param array $items
+   *   The elements as render array.
+   *
+   * @return array
+   *   Render array.
+   */
+  protected function buildCards(array $items): array {
+    return [
+      '#theme' => 'server_theme_cards',
+      '#items' => $items,
+    ];
+  }
+
+  /**
+   * Build an Info cards element.
+   *
+   * @param string $title
+   *   The title.
+   * @param array $body
+   *   The body render array.
+   * @param array $items
+   *   The render array built with
+   *   `ElementLayoutTrait::buildElementLayoutTitleBodyAndItems`.
+   *
+   * @return array
+   *   The render array.
+   */
+  protected function buildElementInfoCards(string $title, array $body, array $items): array {
+    return $this->buildElementLayoutTitleBodyAndItems(
+      $title,
+      $body,
+      $this->buildCards($items),
+    );
+  }
+
+  /**
+   * Build People cards element.
+   *
+   * @param string $title
+   *   The title.
+   * @param array $body
+   *   The body render array.
+   * @param array $items
+   *   The render array built with
+   *   `ElementLayoutTrait::buildElementLayoutTitleBodyAndItems`.
+   *
+   * @return array
+   *   The render array.
+   */
+  protected function buildElementPeopleCards(string $title, array $body, array $items): array {
+    return $this->buildElementLayoutTitleBodyAndItems(
+      $title,
+      $body,
+      $this->buildCards($items),
+    );
+  }
+
+  /**
+   * Build quick links cards.
+   *
+   * @param string $title
+   *   The title.
+   * @param array $body
+   *   The body render array.
+   * @param array $items
+   *   The render array built with
+   *   `ElementLayoutTrait::buildElementLayoutTitleBodyAndItems`.
+   *
+   * @return array
+   *   Render array.
+   */
+  protected function buildElementQuickLinksCards(string $title, array $body, array $items): array {
+    return $this->buildElementLayoutTitleBodyAndItems(
+      $title,
+      $body,
+      $this->buildCards($items),
+    );
+  }
+}
