+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/MediaDocument.php
@@ -5,6 +5,7 @@ namespace Drupal\server_general\Plugin\EntityViewBuilder;
 use Drupal\media\MediaInterface;
 use Drupal\pluggable_entity_view_builder\BuildFieldTrait;
 use Drupal\pluggable_entity_view_builder\EntityViewBuilderPluginAbstract;
+use Drupal\server_general\ElementTrait\DocumentsTrait;
 use Drupal\server_general\InnerElementTrait;
 
 /**
@@ -18,8 +19,7 @@ use Drupal\server_general\InnerElementTrait;
  */
 class MediaDocument extends EntityViewBuilderPluginAbstract {
 
-  use BuildFieldTrait;
-  use InnerElementTrait;
+  use DocumentsTrait;
 
   /**
    * Build the "Card" view mode.
@@ -36,7 +36,7 @@ class MediaDocument extends EntityViewBuilderPluginAbstract {
     /** @var \Drupal\file\FileInterface $file */
     $file = $this->getReferencedEntityFromField($entity, 'field_media_file');
 
-    $build[] = $this->buildInnerElementMediaDocument(
+    $build[] = $this->buildElementDocument(
       $entity->getName(),
       $file->createFileUrl(),
     );
