+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphInfoCards.php
@@ -23,6 +23,7 @@ use Drupal\server_general\ProcessedTextBuilderTrait;
  */
 class ParagraphInfoCards extends EntityViewBuilderPluginAbstract {
 
+  use ElementTrait\CardTrait;
   use ElementTrait;
   use ElementWrapTrait;
   use ProcessedTextBuilderTrait;
