+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphPeopleTeasers.php
@@ -23,6 +23,7 @@ use Drupal\server_general\ProcessedTextBuilderTrait;
  */
 class ParagraphPeopleTeasers extends EntityViewBuilderPluginAbstract {
 
+  use ElementTrait\CardTrait;
   use ElementTrait;
   use ElementWrapTrait;
   use ProcessedTextBuilderTrait;
@@ -43,7 +44,7 @@ class ParagraphPeopleTeasers extends EntityViewBuilderPluginAbstract {
     $paragraphs = $entity->get('field_person_teasers');
     $items = $this->buildReferencedEntities($paragraphs, 'full', $entity->language()->getId());
 
-    $element = $this->buildElementPeopleTeasers(
+    $element = $this->buildElementPeopleCards(
       $this->getTextFieldValue($entity, 'field_title'),
       $this->buildProcessedText($entity, 'field_body'),
       $items,
