+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphText.php
@@ -4,6 +4,7 @@ namespace Drupal\server_general\Plugin\EntityViewBuilder;
 
 use Drupal\paragraphs\ParagraphInterface;
 use Drupal\pluggable_entity_view_builder\EntityViewBuilderPluginAbstract;
+use Drupal\server_general\ElementLayoutTrait;
 use Drupal\server_general\ElementTrait;
 use Drupal\server_general\ProcessedTextBuilderTrait;
 
@@ -19,7 +20,7 @@ use Drupal\server_general\ProcessedTextBuilderTrait;
 class ParagraphText extends EntityViewBuilderPluginAbstract {
 
   use ProcessedTextBuilderTrait;
-  use ElementTrait;
+  use ElementLayoutTrait;
 
   /**
    * Build full view mode.
@@ -33,7 +34,7 @@ class ParagraphText extends EntityViewBuilderPluginAbstract {
    *   Render array.
    */
   public function buildFull(array $build, ParagraphInterface $entity): array {
-    $element = $this->buildElementParagraphTitleAndText(
+    $element = $this->buildElementLayoutTitleAndContent(
       $this->getTextFieldValue($entity, 'field_title'),
       $this->buildProcessedText($entity, 'field_body'),
     );
